feat: 1. add GlobalNotify component 2. use pubsub to call notify event
