+++ src/app/containers/GlobalNotify/Loadable.ts
@@ -1,0 +1,12 @@
+/**
+ *
+ * Asynchronously loads the component for GlobalNotify
+ *
+ */
+
+import { lazyLoad } from 'utils/loadable';
+
+export const GlobalNotify = lazyLoad(
+  () => import('./index'),
+  module => module.GlobalNotify,
+);
