+++ src/locales/zh_cn/translation.json
@@ -253,21 +253,24 @@
       "gte10000": "共找到 {{realTotal}} 个交易（仅展示最新 {{total}} 条记录）"
     },
     "error": {
-      "10001": "列表参数到达最大值。",
-      "20000": "未知错误。",
-      "20001": "响应数据解析错误。",
-      "20002": "请求超时。",
-      "20003": "请求中止。",
-      "400": "错误请求。",
-      "401": "未授权。",
-      "403": "服务器拒绝请求。",
-      "404": "未找到资源。",
-      "405": "方法禁用。",
-      "500": "服务器内部错误。",
-      "501": "服务器不支持请求的功能。",
-      "502": "网关错误。",
-      "503": "服务不可用。",
-      "504": "网关超时。"
+      "title": "系统错误",
+      "description": {
+        "10001": "列表参数到达最大值。",
+        "20000": "未知错误。",
+        "20001": "响应数据解析错误。",
+        "20002": "请求超时。",
+        "20003": "请求中止。",
+        "400": "错误请求。",
+        "401": "未授权。",
+        "403": "服务器拒绝请求。",
+        "404": "未找到资源。",
+        "405": "方法禁用。",
+        "500": "服务器内部错误。",
+        "501": "服务器不支持请求的功能。",
+        "502": "网关错误。",
+        "503": "服务不可用。",
+        "504": "网关超时。"
+      }
     }
   },
   "sponsor": {
